풀이완료 Intermediate_Coder/분할정복/SquareRoot

+++ Beginner_Coder/문자열/VerticalRead.cpp
@@ -68,7 +68,7 @@ void VerticalRead::Code()
 	}
 
 	string result;
-	int idx{ 0 };
+	size_t idx{ 0 };
 	bool isDone{ false };
 
 	while (!isDone)
