+++ Language_Coder/파일입출력/형성평가/FormativeFileProcess10.cpp
@@ -51,7 +51,7 @@ void FormativeFileProcess10::Code()
 		}
 	}
 
-	for (int i = 1; i < results.size(); i++)
+	for (size_t i = 1; i < results.size(); i++)
 	{
 		for (size_t j = 0; j < results.size() - i; j++)
 		{
@@ -64,7 +64,7 @@ void FormativeFileProcess10::Code()
 		}
 	}
 
-	for (int i = 0; i < results.size(); i++)
+	for (size_t i = 0; i < results.size(); i++)
 	{
 		std::cout << results[i] << '\n';
 	}
